refactor VERSION management following example in rapidsai/rmm#1347
