--- python/cucim/VERSION
@@ -1,1 +1,0 @@
-23.12.00
