+++ python/cucim/src/cucim/VERSION
@@ -1,0 +1,1 @@
+../../../../VERSION
